refactor(unified): split ResourceClient interface
