+++ pkg/registry/apis/folders/register_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apiserver/pkg/admission"
 
@@ -712,72 +711,4 @@ func TestFolderAPIBuilder_Mutate_Update(t *testing.T) {
 			require.Equal(t, tt.input, tt.expected)
 		})
 	}
-}
-
-type mockSearchClient struct {
-	mock.Mock
-	resourcepb.ResourceIndexClient
-	resourcepb.ManagedObjectIndexClient
-}
-
-func newMockSearchClient(t *testing.T) *mockSearchClient {
-	t.Helper()
-	searchClient := &mockSearchClient{}
-	t.Cleanup(func() {
-		searchClient.AssertExpectations(t)
-	})
-	return searchClient
-}
-
-func (m *mockSearchClient) Search(ctx context.Context, in *resourcepb.ResourceSearchRequest, opts ...grpc.CallOption) (*resourcepb.ResourceSearchResponse, error) {
-	ret := m.Called(callArgs(ctx, in, opts)...)
-	var resp *resourcepb.ResourceSearchResponse
-	if ret.Get(0) != nil {
-		resp = ret.Get(0).(*resourcepb.ResourceSearchResponse)
-	}
-	return resp, ret.Error(1)
-}
-
-func (m *mockSearchClient) GetStats(ctx context.Context, in *resourcepb.ResourceStatsRequest, opts ...grpc.CallOption) (*resourcepb.ResourceStatsResponse, error) {
-	ret := m.Called(callArgs(ctx, in, opts)...)
-	var resp *resourcepb.ResourceStatsResponse
-	if ret.Get(0) != nil {
-		resp = ret.Get(0).(*resourcepb.ResourceStatsResponse)
-	}
-	return resp, ret.Error(1)
-}
-
-func (m *mockSearchClient) RebuildIndexes(ctx context.Context, in *resourcepb.RebuildIndexesRequest, opts ...grpc.CallOption) (*resourcepb.RebuildIndexesResponse, error) {
-	ret := m.Called(callArgs(ctx, in, opts)...)
-	var resp *resourcepb.RebuildIndexesResponse
-	if ret.Get(0) != nil {
-		resp = ret.Get(0).(*resourcepb.RebuildIndexesResponse)
-	}
-	return resp, ret.Error(1)
-}
-
-func (m *mockSearchClient) CountManagedObjects(ctx context.Context, in *resourcepb.CountManagedObjectsRequest, opts ...grpc.CallOption) (*resourcepb.CountManagedObjectsResponse, error) {
-	ret := m.Called(callArgs(ctx, in, opts)...)
-	var resp *resourcepb.CountManagedObjectsResponse
-	if ret.Get(0) != nil {
-		resp = ret.Get(0).(*resourcepb.CountManagedObjectsResponse)
-	}
-	return resp, ret.Error(1)
-}
-
-func (m *mockSearchClient) ListManagedObjects(ctx context.Context, in *resourcepb.ListManagedObjectsRequest, opts ...grpc.CallOption) (*resourcepb.ListManagedObjectsResponse, error) {
-	ret := m.Called(callArgs(ctx, in, opts)...)
-	var resp *resourcepb.ListManagedObjectsResponse
-	if ret.Get(0) != nil {
-		resp = ret.Get(0).(*resourcepb.ListManagedObjectsResponse)
-	}
-	return resp, ret.Error(1)
-}
-
-func callArgs(ctx context.Context, in interface{}, opts []grpc.CallOption) []interface{} {
-	args := []interface{}{ctx, in}
-	for _, opt := range opts {
-		args = append(args, opt)
-	}
-	return args
 }
