+++ pkg/server/wireexts_oss.go
@@ -147,9 +147,10 @@ var wireExtsBasicSet = wire.NewSet(
 	wire.Bind(new(sandbox.Sandbox), new(*sandbox.Service)),
 	wire.Struct(new(unified.Options), "*"),
 	sql.ProvideStorageBackend,
-	unified.ProvideUnifiedResourceClients,
-	wire.FieldsOf(new(*resource.Client), "StorageClient", "SearchClient", "QuotaClient", "DiagnosticsClient"),
+	unified.ProvideUnifiedResourceClient,
+	wire.FieldsOf(new(*resource.Client), "StorageClient", "SearchClient", "QuotaClient"),
 	wire.Bind(new(resource.MigratorClient), new(*resource.Client)),
+	wire.Bind(new(resource.DiagnosticsClient), new(*resource.Client)),
 	builder.ProvideDefaultBuildHandlerChainFuncFromBuilders,
 	aggregatorrunner.ProvideNoopAggregatorConfigurator,
 	apisregistry.WireSetExts,
